+++ packages/flowr_arch/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.2 2025-8-26
 * adp Dart 3.7
+* feat IRepo::deleteAll; FrRepo::deleteAll;
+* fix FrRepo::update
 
 ## 0.1.1 2025-8-25
 * adp Dart 3.6